linux_requirements.txt
pyfiglet==1.0.3
requests==2.32.4
termcolor==3.1.0
rarfile==4.1
colorama==0.4.6
pycryptodome==3.20.0
bcrypt==4.1.3
passlib==1.7.4
gmssl==3.2.2
paramiko==3.5.1
argon2-cffi==23.1.0
tqdm==4.67.1
